Moved themes to App.tsx and updated bg image
