+++ client/src/containers/AuthContainer.css
@@ -1,8 +1,30 @@
 .AuthContainer {
-  background: url("https://cdn.skyrisecities.com/sites/default/files/images/projects/6977/6977-78314.jpeg");
+  background: url("../assets/Kaneff_Innovation_Complex_Rotunda.jpg");
   width: 100vw;
   height: 100vh;
   background-repeat: no-repeat;
   background-size: cover;
   background-position: center;
+}
+
+.glass::before {
+  content: "";
+	position: absolute;
+	background: inherit;
+	z-index: -1;
+	top: 0;
+	left: 0;
+	right: 0;
+	bottom: 0;
+	box-shadow: inset 0 0 2000px rgba(255, 255, 255, .5);
+	filter: blur(10px);
+	margin: -20px;
+}
+
+.glass {
+	box-shadow: 0 0 1rem 0 rgba(0, 0, 0, .2);	
+	position: relative;
+	z-index: 1;
+	background: inherit;
+	overflow: hidden;
 }
